Release secrets managements, versioning, improved configuration etc

+++ weeve-agent-uninstaller.sh
@@ -9,20 +9,27 @@ log Detecting the OS of the machine ...
 OS=$(uname -s)
 log Detected OS: "$OS"
 
+SERVICE_FILE=/lib/systemd/system/weeve-agent.service
+
+# Exctrating the command to run weeve-agent
+LINE=$(grep "ExecStart" "$SERVICE_FILE")
+COMMAND="sudo ${LINE#ExecStart=} --delete"
+
+WEEVE_AGENT_DIR="$PWD/weeve-agent"
+if [ ! -d "$WEEVE_AGENT_DIR" ]; then
+  log weeve-agent directory does not exists in the current path
+  log please run the script in the path where weeve-agent directory exists
+  exit 1
+fi
+
 if [ "$OS" = "Linux" ]; then
   # if in case the user have deleted the weeve-agent.service and did not reload the systemd daemon
   sudo systemctl daemon-reload
 fi
 
-WEEVE_AGENT_DIR="$PWD/weeve-agent"
-
-SERVICE_FILE=/lib/systemd/system/weeve-agent.service
-
 if [ "$OS" = "Linux" ]; then
   if RESULT=$(systemctl is-active weeve-agent 2>&1); then
     sudo systemctl stop weeve-agent
-    sudo systemctl daemon-reload
-    log weeve-agent service stopped
   else
     log weeve-agent service not running
   fi
@@ -35,6 +42,13 @@ if [ "$OS" = "Linux" ]; then
   fi
 fi
 
+log weeve-agent disconnecting ...
+if RESULT=$(cd "$WEEVE_AGENT_DIR" && eval "$COMMAND" 2>&1); then
+  log weeve-agent disconnected
+else
+  log Error while restarting weeve-agent for disconnection: "$RESULT"
+fi
+
 if [ -d "$WEEVE_AGENT_DIR" ] ; then
   sudo rm -r "$WEEVE_AGENT_DIR"
   log "$WEEVE_AGENT_DIR" removed
@@ -42,4 +56,4 @@ else
   log "$WEEVE_AGENT_DIR" doesnt exists
 fi
 
-log done
+log Removed weeve-agent contents if any.
